+++ hardpy/pytest_hardpy/utils/dialog_box.py
@@ -2,8 +2,10 @@
 # GNU General Public License v3.0 (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from enum import Enum
-from dataclasses import dataclass
-from typing import Any
+from dataclasses import dataclass, asdict
+from typing import Any, List, Optional, Union
+
+from hardpy.pytest_hardpy.utils.exception import WidgetInfoError
 
 
 class DialogBoxWidgetType(Enum):
@@ -16,16 +18,34 @@ class DialogBoxWidgetType(Enum):
 
 
 @dataclass
+class RadiobuttonInfo:
+    fields: List[str]
+
+    def __post_init__(self):
+        if not self.fields:
+            raise ValueError("RadiobuttonInfo must have at least one field")
+
+
+@dataclass
+class CheckboxInfo:
+    fields: List[str]
+
+    def __post_init__(self):
+        if not self.fields:
+            raise ValueError("CheckboxInfo must have at least one field")
+
+
+@dataclass
 class DialogBoxWidget:
     """Dialog box widget.
 
     Args:
         type (DialogBoxWidgetType): widget type
-        info (dict | None): widget info
+        info (Union[RadiobuttonInfo, CheckboxInfo, None]): widget info
     """
 
     type: DialogBoxWidgetType
-    info: dict | None = None
+    info: Optional[Union[RadiobuttonInfo, CheckboxInfo]] = None
 
 
 @dataclass
@@ -53,26 +73,21 @@ def generate_dialog_box_dict(dialog_box_data: DialogBox) -> dict:
         dict: dialog box dictionary
     """
     if dialog_box_data.widget is None:
-        data_dict = {
-            "title_bar": dialog_box_data.title_bar,
-            "dialog_text": dialog_box_data.dialog_text,
-            "widget": None,
-        }
-    else:
-        data_dict = {
-            "title_bar": dialog_box_data.title_bar,
-            "dialog_text": dialog_box_data.dialog_text,
-            "widget": {
-                "info": dialog_box_data.widget.info,
-                "type": dialog_box_data.widget.type.value,
-            },
-        }
-    return data_dict
+        return asdict(dialog_box_data)
+    _validate_widget_info(dialog_box_data.widget)
+
+    def _enum_to_value(data):
+        def convert_value(obj):
+            if isinstance(obj, Enum):
+                return obj.value
+            return obj
+
+        return {k: convert_value(v) for k, v in data}
+
+    return asdict(dialog_box_data, dict_factory=_enum_to_value)
 
 
-def get_dialog_box_data(
-    input_data: str, widget: DialogBoxWidget | None
-) -> Any:
+def get_dialog_box_data(input_data: str, widget: DialogBoxWidget | None) -> Any:
     """Get the dialog box data in the correct format.
 
     Args:
@@ -81,6 +96,9 @@ def get_dialog_box_data(
 
     Returns:
         Any: Dialog box data in the correct format
+
+    Raises:
+        ValueError: If the widget.type is empty.
     """
     if widget is None:
         return None
@@ -88,13 +106,49 @@ def get_dialog_box_data(
     if widget.type is None:
         raise ValueError("Widget type is `None`, but widget data is not empty")
 
+    dbx_answer = None
+
     match widget.type:
         case DialogBoxWidgetType.NUMERIC_INPUT:
             try:
-                return float(input_data)
+                dbx_answer = float(input_data)
             except ValueError:
-                return None
+                dbx_answer = None
         case DialogBoxWidgetType.TEXT_INPUT:
-            return input_data
+            dbx_answer = input_data
+        case DialogBoxWidgetType.RADIOBUTTON:
+            dbx_answer = input_data
+        case DialogBoxWidgetType.CHECKBOX:
+            dbx_answer = input_data
         case _:
-            return None
+            pass  # noqa: WPS420
+
+    return dbx_answer
+
+
+def _validate_widget_info(widget: DialogBoxWidget) -> None:
+    """
+    Validate the information associated with the given widget.
+
+    Args:
+        widget (DialogBoxWidget): The widget to validate.
+
+    Raises:
+        WidgetInfoError: If the widget type is not supported or the info object
+            is not of the expected type for the widget type.
+    """
+    match widget.type:
+        case DialogBoxWidgetType.TEXT_INPUT:
+            if widget.info:
+                raise WidgetInfoError("Expected None for widget info")
+        case DialogBoxWidgetType.NUMERIC_INPUT:
+            if widget.info:
+                raise WidgetInfoError("Expected None for widget info")
+        case DialogBoxWidgetType.RADIOBUTTON:
+            if not isinstance(widget.info, RadiobuttonInfo):
+                raise WidgetInfoError("Expected RadiobuttonInfo for widget info")
+        case DialogBoxWidgetType.CHECKBOX:
+            if not isinstance(widget.info, CheckboxInfo):
+                raise WidgetInfoError("Expected CheckboxInfo for widget info")
+        case _:
+            raise WidgetInfoError(f"Unsupported widget type: {widget.type}")