+++ docs/changelog.md
@@ -3,6 +3,7 @@
 Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<patch>`.
 
 * Fix capturing exceptions and displaying them in the operator panel.
+* Add dialog box with radiobutton and checkbox.
 
 ## HardPy 0.4.0